+++ packages/article-topics/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.8"></a>
+## [1.0.8](https://github.com/newsuk/times-components/compare/@times-components/article-topics@1.0.7...@times-components/article-topics@1.0.8) (2018-05-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-topics
+
 <a name="1.0.7"></a>
 ## [1.0.7](https://github.com/newsuk/times-components/compare/@times-components/article-topics@1.0.6...@times-components/article-topics@1.0.7) (2018-05-21)