+++ packages/watermark/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.5.13"></a>
+## [0.5.13](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.5.12...@times-components/watermark@0.5.13) (2018-05-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/watermark
+
 <a name="0.5.12"></a>
 ## [0.5.12](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.5.11...@times-components/watermark@0.5.12) (2018-05-21)